fix: 문서 디자인 수정
