+++ package.json
@@ -13,7 +13,8 @@
     "axios": "^1.5.0",
     "eslint": "8.49.0",
     "eslint-config-next": "13.4.19",
-    "next": "^13.4.19",
+    "next": "^13.5.6",
+    "next-images": "^1.8.5",
     "next-seo": "^6.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
